Move task 1 to main folder of homework, as main exercise
