+++ src/main/java/geekbrains/homeworks/homework2/App.java
@@ -1,16 +1,13 @@
-package geekbrains.homeworks.homework2.task1;
+package geekbrains.homeworks.homework2;
 
 public class App {
     public static void main(String[] args) {
         Robot robot = new Robot();
-        Animal animal = new Animal();
         Cat cat = new Cat();
         Person person = new Person();
 
         robot.jump();
         robot.run();
-        animal.jump();
-        animal.run();
         cat.jump();
         cat.run();
         person.jump();
