+++ src/main/java/geekbrains/homeworks/homework2/Movement.java
@@ -1,0 +1,7 @@
+package geekbrains.homeworks.homework2;
+
+public interface Movement {
+    void run();
+
+    void jump();
+}
