+++ src/main/java/geekbrains/homeworks/homework2/README.md
@@ -1,0 +1,30 @@
+# Task
+
+## Part 1
+
+Create three classes:
+
+- Person
+- Cat
+- Robot
+
+Which are not inherited from the same class.
+
+These classes should be able to:
+
+- run
+- jump
+
+All with the output of information about the action to the console.
+
+## Part 2
+
+Create two classes:
+
+- Treadmill
+- Wall
+
+When passing through which the participants must perform the appropriate actions
+`run` or `jump`, the result of the execution is printed into the console (successfully
+ran, could not run, etc.). Obstacles have a length of (for the track) or height
+(for the wall), as well as restrictions on running and jumping.
