--- src/main/java/geekbrains/homeworks/homework2/task1/Animal.java
@@ -1,11 +1,0 @@
-package geekbrains.homeworks.homework2.task1;
-
-public class Animal {
-    public void run() {
-        System.out.println("Animal runs");
-    }
-
-    public void jump() {
-        System.out.println("Animal jumps");
-    }
-}
