--- src/main/java/geekbrains/homeworks/homework2/task1/README.md
@@ -1,16 +1,0 @@
-# Task 1
-
-Create three classes:
-
-- Person
-- Cat
-- Robot
-
-Which are not inherited from the same class.
-
-These classes should be able to:
-
-- run
-- jump
-
-All with the output of information about the action to the console.
